moved web pack and karma config files
