+++ config/karma.conf.ci.js
