+++ config/karma.conf.js
@@ -22,8 +22,8 @@ module.exports = function (config) {
 
         // list of files / patterns to load in the browser
         files: [
-            'node_modules/react/dist/react.js',
-            'app/tests/**/*.spec.js',
+            '../node_modules/react/dist/react.js',
+            '../app/tests/**/*.spec.js',
         ],
 
 
@@ -34,7 +34,7 @@ module.exports = function (config) {
         // preprocess matching files before serving them to the browser
         // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
         preprocessors: {
-            'app/tests/**/*.spec.js': ['webpack'],
+            '../app/tests/**/*.spec.js': ['webpack'],
         },
 
         webpack: webpackConfig,
