+++ config/webpack.config.guide.js
