+++ package.json
@@ -4,14 +4,14 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "./node_modules/nodemon/bin/nodemon.js server.js",
-    "guide": "./node_modules/nodemon/bin/nodemon.js server.js --guide",
-    "build": "node ./node_modules/webpack/bin/webpack --config webpack.dist.js",
+    "start": "./node_modules/nodemon/bin/nodemon.js index.js",
+    "guide": "./node_modules/nodemon/bin/nodemon.js index.js --guide",
+    "build": "node ./node_modules/webpack/bin/webpack --config config/webpack.dist.js",
     "app": "npm run build && ./node_modules/http-server/bin/http-server ./dist -p 3000",
     "lint": "./node_modules/eslint/bin/eslint.js ./",
     "setup-ci": "npm install --save-dev chai@^3.4.x karma@^0.13.x karma-chai@^0.1.x karma-es5-shim@0.0.x karma-mocha@^0.2.x karma-phantomjs-launcher@^0.2.x karma-sinon@^1.0.x karma-webpack@^1.7.x mocha@^2.3.x phantomjs@^1.9.x sinon@^1.17.x",
-    "ci": "./node_modules/karma/bin/karma start ./karma.conf.ci.js",
-    "test": "./node_modules/karma/bin/karma start ./karma.conf.js"
+    "ci": "./node_modules/karma/bin/karma start ./config/karma.conf.ci.js",
+    "test": "./node_modules/karma/bin/karma start ./config/karma.conf.js"
   },
   "author": "",
   "license": "ISC",
